CHANGES 10-16-2021
Removed PhysicalDrives. I am trying to re-approach adding a file permissions system to the game and having to differentiate between different drives on a system is too much for me. As such, there are no more "Secrets Drives" or other drives of any type.

Conversely, this will make it way easier to add in a permissions system and will also make the structure of computers in the game a bit more faithful to real life.

The game does not boot right now, still working on that.