CHANGES 02-12-2022
Finished almost all of what was left to do for the new permissions system. The new permissions are as follows:

For directories:
E - Enter (perform CD command)
L - List (perform LS command)
M - Modify (add, remove, or delete children)

For files:
R - Read (get contents of file)
W - Write (modify contents of file)
E - Execute (attempt to invoke file as an executable)

Only thing left to do is update SecureCopy which I am expecting to be a pain in the ass so that is on the backburner for now.

Also, yes, this took longer than it should have.
I am lazy.